try as embedding layer

+++ src/data.py
@@ -247,7 +247,7 @@ class TreeDataset(Dataset):
             
         if self.metadata:
             site = self.annotations.site.loc[index]  
-            site = torch.tensor(site, dtype=torch.float)
+            site = torch.tensor(site, dtype=torch.int)
             site = torch.unsqueeze(site, dim=0)
             inputs["site"] = site
